feat: implement initial database migration and model setup for team planning
- Added initial migration to create database schema for Boards, Features, Sprints, TeamMembers, UserStories, and TeamMemberSprints.
- Updated model snapshot to reflect new database structure.
- Modified TeamMemberSprint model to change CapacityDev and CapacityTest types from double to int.
- Implemented methods in TeamRepository and ITeamRepository for deleting team members.
- Enhanced TeamService to support adding, updating, and deleting team members with capacity management.
- Updated API endpoints and service methods for team member management.
- Improved frontend components for team member actions, including edit and delete functionalities.
- Added validation for team member name and capacity inputs in the UI.

+++ backend/pi-planning-backend/Migrations/20260215065624_InitialCreate.cs
@@ -147,8 +147,8 @@ namespace PiPlanningBackend.Migrations
                         .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
                     TeamMemberId = table.Column<int>(type: "integer", nullable: false),
                     SprintId = table.Column<int>(type: "integer", nullable: false),
-                    CapacityDev = table.Column<double>(type: "float", nullable: false),
-                    CapacityTest = table.Column<double>(type: "float", nullable: false)
+                    CapacityDev = table.Column<int>(type: "integer", nullable: false),
+                    CapacityTest = table.Column<int>(type: "integer", nullable: false)
                 },
                 constraints: table =>
                 {
